# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Church OS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface SignupRequest {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  phoneNumber?: string;
}

export interface AuthUser {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  preferredName: string | null;
  /** @nullable */
  phoneNumber: string | null;
  /** @nullable */
  profilePhotoUrl: string | null;
  /** @nullable */
  dateOfBirth: string | null;
  /** @nullable */
  gender: string | null;
  /** @nullable */
  maritalStatus: string | null;
  /** @nullable */
  occupation: string | null;
  /** @nullable */
  preferredLanguage: string | null;
  /** @nullable */
  emergencyContactName: string | null;
  /** @nullable */
  emergencyContactPhoneNumber: string | null;
  /** @nullable */
  streetAddress: string | null;
  /** @nullable */
  apartmentUnit: string | null;
  /** @nullable */
  city: string | null;
  /** @nullable */
  state: string | null;
  /** @nullable */
  zipCode: string | null;
  /** @nullable */
  country: string | null;
  role: string;
  /** @nullable */
  adminLevel: string | null;
  /** @nullable */
  adminTitle: string | null;
  /** @nullable */
  assignedMinistry: string | null;
  accountStatus: string;
  /** @nullable */
  createdByUserId: number | null;
  churchId: number;
  churchName: string;
  createdAt: string;
  /** @nullable */
  lastLoginAt: string | null;
  authProviders: string[];
  hasPassword: boolean;
  adminPermissions: string[];
}

export interface UpdateProfileRequest {
  firstName: string;
  lastName: string;
  /** @nullable */
  preferredName?: string | null;
  email: string;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  profilePhotoUrl?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  maritalStatus?: string | null;
  /** @nullable */
  occupation?: string | null;
  /** @nullable */
  preferredLanguage?: string | null;
  /** @nullable */
  emergencyContactName?: string | null;
  /** @nullable */
  emergencyContactPhoneNumber?: string | null;
  /** @nullable */
  streetAddress?: string | null;
  /** @nullable */
  apartmentUnit?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zipCode?: string | null;
  /** @nullable */
  country?: string | null;
}

export interface ChangePasswordRequest {
  currentPassword: string;
  newPassword: string;
}

export interface OAuthConfig {
  google: boolean;
  apple: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}