# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Church OS API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      description: Authenticate with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Signup
      description: Create a new member account and begin a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Signup successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      description: Returns the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update member profile
      description: Update the authenticated user's profile details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      description: Change the authenticated user's password when password auth is enabled
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Password auth unavailable or invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated or current password invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      description: Clear the current session
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/oauth/config:
    get:
      operationId: getOAuthConfig
      tags: [auth]
      summary: Get OAuth provider availability
      responses:
        "200":
          description: OAuth config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthConfig"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    SignupRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        phoneNumber:
          type: string
      required:
        - firstName
        - lastName
        - email
        - password

    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        preferredName:
          type:
            - string
            - "null"
        phoneNumber:
          type:
            - string
            - "null"
        profilePhotoUrl:
          type:
            - string
            - "null"
        dateOfBirth:
          type:
            - string
            - "null"
        gender:
          type:
            - string
            - "null"
        maritalStatus:
          type:
            - string
            - "null"
        occupation:
          type:
            - string
            - "null"
        preferredLanguage:
          type:
            - string
            - "null"
        emergencyContactName:
          type:
            - string
            - "null"
        emergencyContactPhoneNumber:
          type:
            - string
            - "null"
        streetAddress:
          type:
            - string
            - "null"
        apartmentUnit:
          type:
            - string
            - "null"
        city:
          type:
            - string
            - "null"
        state:
          type:
            - string
            - "null"
        zipCode:
          type:
            - string
            - "null"
        country:
          type:
            - string
            - "null"
        role:
          type: string
        adminLevel:
          type:
            - string
            - "null"
        adminTitle:
          type:
            - string
            - "null"
        assignedMinistry:
          type:
            - string
            - "null"
        accountStatus:
          type: string
        createdByUserId:
          type:
            - integer
            - "null"
        churchId:
          type: integer
        churchName:
          type: string
        createdAt:
          type: string
        lastLoginAt:
          type:
            - string
            - "null"
        authProviders:
          type: array
          items:
            type: string
        hasPassword:
          type: boolean
        adminPermissions:
          type: array
          items:
            type: string
      required:
        - id
        - email
        - firstName
        - lastName
        - preferredName
        - phoneNumber
        - profilePhotoUrl
        - dateOfBirth
        - gender
        - maritalStatus
        - occupation
        - preferredLanguage
        - emergencyContactName
        - emergencyContactPhoneNumber
        - streetAddress
        - apartmentUnit
        - city
        - state
        - zipCode
        - country
        - role
        - adminLevel
        - adminTitle
        - assignedMinistry
        - accountStatus
        - createdByUserId
        - churchId
        - churchName
        - createdAt
        - lastLoginAt
        - authProviders
        - hasPassword
        - adminPermissions

    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        preferredName:
          type:
            - string
            - "null"
        email:
          type: string
        phoneNumber:
          type:
            - string
            - "null"
        profilePhotoUrl:
          type:
            - string
            - "null"
        dateOfBirth:
          type:
            - string
            - "null"
        gender:
          type:
            - string
            - "null"
        maritalStatus:
          type:
            - string
            - "null"
        occupation:
          type:
            - string
            - "null"
        preferredLanguage:
          type:
            - string
            - "null"
        emergencyContactName:
          type:
            - string
            - "null"
        emergencyContactPhoneNumber:
          type:
            - string
            - "null"
        streetAddress:
          type:
            - string
            - "null"
        apartmentUnit:
          type:
            - string
            - "null"
        city:
          type:
            - string
            - "null"
        state:
          type:
            - string
            - "null"
        zipCode:
          type:
            - string
            - "null"
        country:
          type:
            - string
            - "null"
      required:
        - firstName
        - lastName
        - email

    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword

    OAuthConfig:
      type: object
      properties:
        google:
          type: boolean
        apple:
          type: boolean
      required:
        - google
        - apple

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message