# lib/api-zod/src/generated/types/oAuthConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Church OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OAuthConfig {
  google: boolean;
  apple: boolean;
}